requirements.txt
asgiref==3.4.1
Django==3.2.7
numpy==1.21.2
Pillow==8.3.2
pytz==2021.1
sqlparse==0.4.1
